+++ src/ewolucyjne/Punkt.java
@@ -5,37 +5,37 @@ import java.util.Random;
 
 public class Punkt implements Osobnik {
 
-	private float wspX;
-	private float wspY;
-	private float sigmaX;
-	private float sigmaY;
-	private float warFun;
+	private double wspX;
+	private double wspY;
+	private double sigmaX;
+	private double sigmaY;
+	private double warFun;
 	private FunkcjaPrzystosowania func;
 	Punkt( )
 	{}
 	
 	@Override
-	public void inicjowanie( ArrayList<Float> parametry, ArrayList<Float> sigmy , FunkcjaPrzystosowania fun)
+	public void inicjowanie( ArrayList<Double> parametry, ArrayList<Double> sigmy , FunkcjaPrzystosowania fun)
 	{
 		this.wspX = parametry.get(0);
 		this.wspY = parametry.get(1);
 		this.sigmaX = sigmy.get(0);
 		this.sigmaY = sigmy.get(1);
 		
-		ArrayList<Float> tmp = new ArrayList<Float>();
+		ArrayList<Double> tmp = new ArrayList<Double>();
 		tmp.add(wspX);
 		tmp.add(wspY);
 		this.warFun = fun.Oblicz(tmp);
 		this.func = fun;
 	}
 	@Override
-	public float pobierzWartosc() 
+	public double pobierzWartosc() 
 	{
 		return warFun;
 	}
 
 	@Override
-	public float pobierzParametr(int n) 
+	public double pobierzParametr(int n) 
 	{
 		if( n == 0 )
 			return wspX;
@@ -49,7 +49,7 @@ public class Punkt implements Osobnik {
 	}
 
 	@Override
-	public float pobierzSigme(int n) 
+	public double pobierzSigme(int n) 
 	{
 		if( n == 0 )
 			return sigmaX;
@@ -72,7 +72,7 @@ public class Punkt implements Osobnik {
 			sigmaY *= Math.exp(t1*N + t*gen.nextGaussian());
 			wspX += sigmaX * gen.nextGaussian();
 			wspY += sigmaY * gen.nextGaussian();
-			ArrayList<Float> tmp = new ArrayList<Float>();
+			ArrayList<Double> tmp = new ArrayList<Double>();
 			tmp.add(wspX);
 			tmp.add(wspY);
 			this.warFun = this.func.Oblicz(tmp);
